[SE-265] Remove set_addresses for Node object
Setting addresses on the Node level may cause errors, because it supposes
setting single node address. But if second address is not set (e.g. on a data
port), tatlin will raise an error. Besides two different possibilities for the
same action can mislead user. So it was decided to leave setting address only
on port level