Object checker can accept lists of objects + add ignoring order in lists
Checker processed only one object at once. It was not convenient for checking
multiple objects. Now checker also accepts a list of objects. Also, it is
the reason to rename it. check_object and check_objects can mislead the one,
because it can check both.

Ignoring order is useful for Python 2, when lists are retrieved from dicts,
which have no order in Python 2